feat: implement ChangePasswordCard component for password updates
- Added a new ChangePasswordCard component to encapsulate the password change functionality, improving code organization and reusability.
- Integrated form handling with validation using Zod and React Hook Form, including error handling and success messages.
- Updated the account security route to utilize the new ChangePasswordCard component, streamlining the user interface for password management.
